adds readme and fixes import task

+++ db/migrate/20181127202017_create_invoice_items.rb
@@ -4,7 +4,7 @@ class CreateInvoiceItems < ActiveRecord::Migration[5.2]
       t.references :item, foreign_key: true
       t.references :invoice, foreign_key: true
       t.integer :quantity
-      t.integer :unit_price
+      t.bigint :unit_price
 
       t.timestamps
     end
